git-test-sequence: quieter, friendlier output
The current handling of test output does not scale well.  Currently,
since the test output is sent to the terminal, it can be quite
difficult to tell what has failed until the overall report at the end,
and by the end report the failure details could be long gone or at
least difficult to find.

Instead, send the output to log files named by hash. This frees us up
to have much nicer output, so for each commit simply show the commit
line followed by a timer while the test is running, and PASS/FAIL when
the test finishes. This makes it easy to judge progress both through
the commits and within the current commit and to see the results of
tests so far.