configure-gnome: rewrite in Python, add PaperWM setup
I finally pushed too hard on the gsettings command line, so this is a
total rewrite in Python that actually understands the structured types
in the gsettings database and uses the GLib APIs directly.

This in turn let me do all of my PaperWM setup. Stronger typing was,
in particular, necessary to manipulate PaperWM key binding settings in
a rational way.